+++ wtfix/apps/parsers.py
@@ -21,7 +21,6 @@ from wtfix.conf import settings, logger
 from wtfix.core.utils import GroupTemplateMixin
 from wtfix.message.field import Field
 from wtfix.message.message import RawMessage, generic_message_factory
-from wtfix.protocol.common import Tag
 
 
 class RawMessageParserApp(BaseApp, GroupTemplateMixin):
@@ -42,12 +41,12 @@ class RawMessageParserApp(BaseApp, GroupTemplateMixin):
         )  # Remove last SOH at end of byte stream and split into fields
 
         fields = [
-            message[Tag.BeginString],
-            message[Tag.BodyLength],
-            message[Tag.MsgType],
-            message[Tag.MsgSeqNum],
+            message.BeginString,
+            message.BodyLength,
+            message.MsgType,
+            message.MsgSeqNum,
             *self._parse_fields(data),
-            message[Tag.CheckSum],
+            message.CheckSum,
         ]
 
         message = generic_message_factory(*fields, group_templates=self.group_templates)